Initial Commit
Created the basic foundation for the game.  Found online (usable) images of chess pieces, created the board using a JFrame and designated x,y positions.  Also created the mediator/listener that creates a bridge between the mouse interaction and the piece object.  The Piece object is a simple object with an Image to see the piece, as well as x,y properties for the position of the upper right position of the piece (the offset variables designate where the piece should follow the mouse's movement.