MVC Refactoring
Thanks to intelliJs awesome Refactor feature, I was able to easily convert the code I had into the useful MVC pattern for development.  I also abstracted many more objects, creating a Board Model and a Piece model, separating that logic from the GameGUI and the PieceView classes which are the actual view classes themselves.  The Mediator still acts as the intermediary between the two (should technically be a controller, but we can deal with that later), and the GameController is the functionality of the game of chess, while the GameGUI is the interaction with the screen.  This allows for more flexibility and multiple applications of the same business logic.