Added Move Validator funcationality
Added a class called MoveValidator in the controller package.  This class is meant to act as a validator for the pieces move -- game logic -- and make sure that the move the user is trying to attempt is valid.  This is the basic foundation, as there is a lot I want to add, but may not be within the scope of this project: checkmate, stalemate, etc.  I probably will add them later, but not as of now.   I also want to abstract the row/column integers into a subclass of the Board model into something like a "Tile" class or along those lines.  Idk, just throwing out ideas