Incorporated the MoveValidator
I incorporated the MoveValidator into the GameController to actually check to make sure the moves are valid and prevent invalid moves from being executed.  I also added a game end gameState, to signify the end of the game.  As of know, the end of the game is signified by the capturing of one of hte kings, but once I add some AI logic or possible move logic I will be able to determine whether a checkmate state has been reached