MAJOR UPDATES
I kinda got lost in the code, so there are a ton of updates since the last commit.  Things like this happen.  Long story short, as you can see in the code, there are a ton of changes to make things more design-pattern oriented.  This is something I learned from working at the startup I worked for Freshman/Sophomore year and have picked up in general.  You will also see thats why there is extenesive commenting.  Basically, I separated the view from the business logic even farther by denoting models for each separate idea.  The MAIN difference is the addition of the network, which I havent created the server for yet.  This required me to add a lot of methods that converted to strings in certain formats, which I will probably add to a README.md at the end of this project.  However, for now, not only does the game logic DRIVE the GUI (as opposed to the way it was before where the user action drove the game logic), but there are listeners and events set up everywhere to update certain actions depending on what happened (from the business logic side).  I will make a more detailed description (though most can be seen in my comments), but for now I need to get away from a computer screen.